added comments and moved getPackageMetadata
