+++ packages/modular-scripts/src/buildPackage/makeBundle.ts
@@ -15,7 +15,7 @@ import resolve from '@rollup/plugin-node-resolve';
 
 import { getLogger } from './getLogger';
 import { getPackageEntryPoints } from './getPackageEntryPoints';
-import getPackageMetadata from './getPackageMetadata';
+import getPackageMetadata from '../utils/getPackageMetadata';
 import getModularRoot from '../utils/getModularRoot';
 
 const outputDirectory = 'dist';
